requirements.txt
selenium==4.16.0
pymongo==4.6.1
flask==3.0.0
python-dotenv==1.0.0